RavenDB-20732 fix tests and refactor
