+++ src/Raven.Server/Documents/DocumentPutAction.cs
@@ -212,7 +212,7 @@ namespace Raven.Server.Documents
                     }
                 }
 
-                var result = BuildChangeVectorAndResolveConflicts(context, lowerId, newEtag, document, changeVector, expectedChangeVector, flags, oldChangeVector);
+                var result = _documentsStorage.BuildChangeVectorAndResolveConflicts(context, lowerId, newEtag, document, changeVector, expectedChangeVector, flags, oldChangeVector);
 
                 nonPersistentFlags |= result.NonPersistentFlags;
 
@@ -318,81 +318,6 @@ namespace Raven.Server.Documents
                 throw new InvalidDataException("The incoming document " + id + " has changed _during_ the put process, " +
                                                "this is likely because you are trying to save a document that is already stored and was moved");
             }
-        }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public (ChangeVector ChangeVector, NonPersistentDocumentFlags NonPersistentFlags) BuildChangeVectorAndResolveConflicts(
-            DocumentsOperationContext context, Slice lowerId, long newEtag,
-            BlittableJsonReaderObject document, ChangeVector changeVector, string expectedChangeVector, DocumentFlags flags, ChangeVector oldChangeVector)
-        {
-            var nonPersistentFlags = NonPersistentDocumentFlags.None;
-            var fromReplication = flags.Contain(DocumentFlags.FromReplication);
-
-            if (_documentsStorage.ConflictsStorage.ConflictsCount != 0)
-            {
-                // Since this document resolve the conflict we don't need to alter the change vector.
-                // This way we avoid another replication back to the source
-
-                _documentsStorage.ConflictsStorage.ThrowConcurrencyExceptionOnConflictIfNeeded(context, lowerId, expectedChangeVector);
-
-                if (fromReplication)
-                {
-                    nonPersistentFlags = _documentsStorage.ConflictsStorage.DeleteConflictsFor(context, lowerId, document).NonPersistentFlags;
-                }
-                else
-                {
-                    var result = _documentsStorage.ConflictsStorage.MergeConflictChangeVectorIfNeededAndDeleteConflicts(changeVector, context, lowerId, newEtag, document);
-                    nonPersistentFlags = result.NonPersistentFlags;
-
-                    if (result.ChangeVector != null)
-                        return (result.ChangeVector, nonPersistentFlags);
-                }
-            }
-
-            if (changeVector != null)
-                return (changeVector, nonPersistentFlags);
-
-            if (fromReplication == false)
-            {
-                context.LastDatabaseChangeVector ??= GetDatabaseChangeVector(context);
-                oldChangeVector = oldChangeVector == null
-                    ? context.GetChangeVector(context.LastDatabaseChangeVector)
-                    : oldChangeVector.MergeWith(context.LastDatabaseChangeVector, context);
-            }
-
-            changeVector = SetDocumentChangeVectorForLocalChange(context, lowerId, oldChangeVector, newEtag);
-            context.SkipChangeVectorValidation = changeVector.TryRemoveIds(_documentsStorage.UnusedDatabaseIds, context, out changeVector);
-            return (changeVector, nonPersistentFlags);
-        }
-
-        public static bool UpdateLastDatabaseChangeVector(DocumentsOperationContext context, ChangeVector changeVector, DocumentFlags flags, NonPersistentDocumentFlags nonPersistentFlags)
-        {
-            // if arrived from replication we keep the document with its original change vector
-            // in that case the updating of the global change vector should happened upper in the stack
-            if (nonPersistentFlags.Contain(NonPersistentDocumentFlags.FromReplication))
-                return false;
-
-            var currentGlobalChangeVector = context.LastDatabaseChangeVector ?? GetDatabaseChangeVector(context);
-
-            var clone = context.GetChangeVector(changeVector);
-            clone = clone.StripSinkTags(currentGlobalChangeVector, context);
-
-            // this is raft created document, so it must contain only the RAFT element 
-            if (flags.Contain(DocumentFlags.FromClusterTransaction))
-            {
-                context.LastDatabaseChangeVector = ChangeVector.Merge(currentGlobalChangeVector, clone.Order, context);
-                return false;
-            }
-
-            // the resolved document must preserve the original change vector (without the global change vector) to avoid ping-pong replication.
-            if (nonPersistentFlags.Contain(NonPersistentDocumentFlags.FromResolver))
-            {
-                context.LastDatabaseChangeVector = ChangeVector.Merge(currentGlobalChangeVector, clone.Order, context);
-                return false;
-            }
-
-            context.LastDatabaseChangeVector = clone.Order;
-            return true;
         }
 
         protected virtual void CalculateSuffixForIdentityPartsSeparator(string id, ref char* idSuffixPtr, ref int idSuffixLength, ref int idLength)
@@ -779,17 +704,6 @@ namespace Raven.Server.Documents
                     return;
                 }
             }
-        }
-
-        private ChangeVector SetDocumentChangeVectorForLocalChange(DocumentsOperationContext context, Slice lowerId, ChangeVector oldChangeVector, long newEtag)
-        {
-            if (oldChangeVector != null)
-            {
-                oldChangeVector = oldChangeVector.UpdateVersion(_documentDatabase.ServerStore.NodeTag, _documentsStorage.Environment.Base64Id, newEtag, context);
-                oldChangeVector = oldChangeVector.UpdateOrder(_documentDatabase.ServerStore.NodeTag, _documentsStorage.Environment.Base64Id, newEtag, context);
-                return oldChangeVector;
-            }
-            return context.GetChangeVector(_documentsStorage.ConflictsStorage.GetMergedConflictChangeVectorsAndDeleteConflicts(context, lowerId, newEtag));
         }
 
         [Conditional("DEBUG")]
