+++ test/Tests.Infrastructure/ReplicationTestBase.cs
@@ -585,5 +585,18 @@ namespace Tests.Infrastructure
                 Result = result;
             }
         }
+
+        public class NoChangeVectorContext : IChangeVectorOperationContext
+        {
+            public static NoChangeVectorContext Instance = new NoChangeVectorContext();
+
+            public ChangeVector GetChangeVector(string changeVector, bool throwOnRecursion = false) => new ChangeVector(changeVector, throwOnRecursion, this);
+
+            public ChangeVector GetChangeVector(string version, string order)
+            {
+                return new ChangeVector(new ChangeVector(version, throwOnRecursion: true, this), 
+                    new ChangeVector(version, throwOnRecursion: true, this));
+            }
+        }
     }
 }
